Extracts the Confidence Intervals Computed for Each Category for the 'Adapted Paik et Al.' Model

Extracts the confidence intervals for \(\boldsymbol{\phi}\), \(\boldsymbol{\beta}\), \(\mu_1\), \(\nu\), \(\boldsymbol{\gamma}\) obtained with the time-dependent frailty model proposed in the 'Adapted Paik et al.' framework.

Usage

# S3 method for class 'AdPaik'
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

An S3 object of class AdPaik, returned by the main model function (AdPaikModel). This object contains all the optimal parameter estimates.

parm

A specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. Defaults to NULL, and all parameters are considered. Changing it is not supported for this model. It will be ignored.

level

The confidence level required. Defaults to 0.95.

...

Additional arguments to be passed to other methods.

Value

A named list containing the categories of the standard errors for the optimal parameters.

Details

The confint.AdPaik function extracts the standard errors for the estimated parameters from the ParametersCI field in object.

The function validates the structure of object and ensures compatibility with the expected model output. It throws an error if the object is malformed or inconsistent.

Examples

# Example using the 'Academic Dropout' dataset
data(data_dropout)

# Define the formula and time axis for the model
formula <- time_to_event ~ Gender + CFUP + cluster(group)
time_axis <- c(1.0, 1.4, 1.8, 2.3, 3.1, 3.8, 4.3, 5.0, 5.5, 5.8, 6.0)
eps <- 1e-10
categories_range_min <- c(-8, -2, eps, eps, eps)
categories_range_max <- c(-eps, 0, 1 - eps, 1, 10)

# \donttest{
# Run the main model
result <- AdPaikModel(formula, data_dropout, time_axis,
                      categories_range_min, categories_range_max, TRUE)

# Extract the coefficients
confint(result)
#> $confint.phi
#> $left
#>  [1] -3.671957 -4.040163 -2.729596 -5.157327 -4.332896 -3.307973 -5.366439
#>  [8] -3.380787 -4.951976 -4.210769
#> 
#> $right
#>  [1] -3.671957 -4.040163 -2.729596 -5.157327 -4.332896 -3.307973 -5.366439
#>  [8] -3.380787 -4.951976 -4.210769
#> 
#> attr(,"class")
#> [1] "phi"
#> 
#> $confint.beta
#> $left
#> GenderMale       CFUP 
#>  0.1027304 -1.2071157 
#> 
#> $right
#> GenderMale       CFUP 
#>  0.1027304 -1.2071157 
#> 
#> attr(,"class")
#> [1] "beta"
#> 
#> $confint.mu1
#> $left
#> [1] 0.009837761
#> 
#> $right
#> [1] 0.009837761
#> 
#> attr(,"class")
#> [1] "mu1"
#> 
#> $confint.nu
#> $left
#> [1] 429.5328
#> 
#> $right
#> [1] 429.5328
#> 
#> attr(,"class")
#> [1] "nu"
#> 
#> $confint.gamma
#> $left
#>  [1] 0.006383573 0.311030038 0.200241483 0.006126961 0.240187739 0.228218767
#>  [7] 1.292007407 0.427801190 0.006091840 0.728564675
#> 
#> $right
#>  [1] 0.006383573 0.311030038 0.200241483 0.006126961 0.240187739 0.228218767
#>  [7] 1.292007407 0.427801190 0.006091840 0.728564675
#> 
#> attr(,"class")
#> [1] "gamma"
#> 
# }