Posterior frailty estimates and variances for the 'Adapted Paik et al.'s Model'

Function for computing the posterior frailty estimates and variances of the time-dependent shared frailty Cox model. Recalling the structure of the frailty \(Z_{jk} = \alpha_j + \epsilon_{jk}, \forall j,k\) as being composed by the sum of two independent gamma distributions:

\(\alpha_j \sim gamma(\mu_1/\nu, 1/\nu), \forall j\)

\(\epsilon_{jk} \sin gamma(\mu_2/\gamma_k, 1/\gamma_k), \forall j,k\) the posterior distribution of both terms is still a gamma with different mean and variance and the posterior frailty estimate corresponds to the 'empirical Bayes estimate', that is the previous mentioned posterior mean.

Usage

post_frailty.AdPaik(optimal_params, dataset, time_to_event, centre, time_axis)

Arguments

optimal_params

Optimal parameters estimated by maximizing the log-likelihood function, through the constraint multi-dimensional optmization method.

dataset

Dataset containing all the covariates/regressors.

time_to_event

Time-instant, in the follow-up, in which an individual faces the event or fails. If an individual does not face the event in the follow-up, then the time-instant must assume a default value.

centre

Individual group/cluster membership.

time_axis

Temporal domain.

Value

S3 object of class 'PF.AdPaik' composed of two elements of different class:

PosteriorFrailtyEst: S3 object of class 'PFE.AdPaik'.

PosteriorFrailtyVar: S3 object of class 'PFV.AdPaik'.