Summary of the Adapted Paik et al.'s Time-Dependent Shared Frailty Model

This function provides a comprehensive summary of the results from the Adapted Paik et al.'s Time-Dependent Shared Frailty Model. It includes key information about the dataset (e.g., number of individuals, regressors, intervals, and clusters), model parameters, and output (log-likelihood, AIC). The summary also lists the estimated regressors along with their standard errors

Usage

# S3 method for class 'AdPaik'
summary(result)

Arguments

result

'S3' class object returned by the main model call, i.e. output of the 'Adapted Paik et al.'s Model'.

Value

Model summary printed on output.

Details

The function reports the estimated regressors, their standard errors, and confidence intervals (if available).

Examples

# Define the variables needed for the model execution

data(data_dropout)
#> Warning: data set ‘data_dropout’ not found
eps_paik <- 1e-10
categories_range_min <- c(-8, -2, eps_paik, eps_paik, eps_paik)
categories_range_max <- c(-eps_paik, 0.4, 1 - eps_paik, 1, 10)
time_axis <- c(1.0, 1.4, 1.8, 2.3, 3.1, 3.8, 4.3, 5.0, 5.5, 5.8, 6.0)
formula <- time_to_event ~ Gender + CFUP + cluster(group)

# \donttest{
# Call the main model function
result <- AdPaikModel(formula, data_dropout, time_axis, categories_range_min, categories_range_max)
#> Error: object 'data_dropout' not found

# Call the summary
summary(result)
#> Error: object 'result' not found
# }