Posterior Frailty Estimates

Function for computing the posterior frailty estimates of the time-dependent shared frailty Cox model.

Recalling the structure of the frailty \(Z_{jk} = \alpha_j + \epsilon_{jk}, \forall j,k\) with \(k=1,\dots,L\) and \(j=1,\dots,N\) as being composed by the sum of two independent gamma distributions:

\(\alpha_j \sim gamma(\mu_1/\nu, 1/\nu), \forall j\)

\(\epsilon_{jk} \sim gamma(\mu_2/\gamma_k, 1/\gamma_k), \forall j,k\)

The posterior frailty estimate is \(\hat{Z}_{jk} = \hat{\alpha}_{j}/\hat{\alpha}_{max} + \hat{\epsilon}_{jk}/\hat{\epsilon}_{max}\). This function allows to get either the entire posterior frailty estimate \(\hat{Z}_{jk}\) or its time-independent \(\frac{\hat{\alpha}_{j}}{\hat{\alpha}_{\text{max}}}\) or time-dependent \(\frac{\hat{\epsilon}_{jk}}{\hat{\epsilon}_{\text{max}}}\) components. The user can control which components to display using the flag_eps and flag_alpha parameters. Only one of these flags can be set to TRUE at a time.

Usage

post_frailty_est(object, flag_eps = FALSE, flag_alpha = FALSE)

Arguments

object

S3 object of class 'AdPaik' returned by the main model output, that contains all the information for the computation of the frailty standard deviation.

flag_eps

Logical flag indicating whether to extract only the time-dependent posterior frailty estimates. Default is FALSE.

flag_alpha

Logical flag indicating whether to extract only the time-independent posterior frailty estimates. Default is FALSE.

Value

Vector or matrix of posterior frailty estimates, depending on the flag_eps and flag_alpha values. Specifically:

It is a vector of length equal to the N containing posterior frailty estimates for \(\alpha_j, \forall j\). In this case the flag_eps must be FALSE and the flag_alpha must be TRUE.

Matrix of dimension (N, L) containing posterior frailty estimates for \(\epsilon_{jk}, \forall j,k\). In this case the flag_eps must be TRUE and the flag_alpha must be FALSE.

Matrix of dimension (N, L) containing posterior frailty estimates for \(Z_{jk} = \alpha_j + \epsilon_{jk}, \forall j,k\). In this case the flag_eps must be FALSE and the flag_alpha must be FALSE.

Examples

# Consider the 'Academic Dropout dataset'
data(data_dropout)

# Define the variables needed for the model execution
formula <- time_to_event ~ Gender + CFUP + cluster(group)
time_axis <- c(1.0, 1.4, 1.8, 2.3, 3.1, 3.8, 4.3, 5.0, 5.5, 5.8, 6.0)
eps <- 1e-10
categories_range_min <- c(-8, -2, eps, eps, eps)
categories_range_max <- c(-eps, 0, 1 - eps, 1, 10)

# \donttest{
# Call the main model
result <- AdPaikModel(formula, data_dropout, time_axis,
                      categories_range_min, categories_range_max)

post_frailty_est(result)
#>            [,1]     [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 1.1048099 1.108778 1.1474606 1.1059507 1.1023955 0.9904692 1.0244158
#>  [2,] 0.8379035 0.783926 0.6487763 0.8449023 0.8444907 0.8022560 0.7888707
#>  [3,] 1.1625687 1.132694 1.1164256 1.1623134 1.1654970 1.1567724 1.1361889
#>  [4,] 1.4271733 1.568561 1.6006542 1.4236236 1.4622341 1.4087053 1.3630470
#>  [5,] 1.0151256 1.010793 0.8952022 1.0126754 1.0077753 0.9719531 0.9590429
#>  [6,] 1.1586254 1.137414 1.2133180 1.1713028 1.1852623 1.0503533 1.1159622
#>  [7,] 1.5066891 1.523512 1.4942505 1.5035662 1.4868939 1.4476778 1.4707300
#>  [8,] 1.1898437 1.165316 1.1408176 1.1802813 1.1676276 1.1581899 1.2267386
#>  [9,] 1.2029324 1.202930 1.1728438 1.2033448 1.2188235 1.1907893 1.1795115
#> [10,] 1.2415568 1.279429 1.2641002 1.2343066 1.2156415 1.2301263 1.4222394
#> [11,] 1.3800678 1.363966 1.5389846 1.3763751 1.3666882 1.3816956 1.3480006
#> [12,] 1.3800415 1.447069 1.5676168 1.3933870 1.4089605 1.6238984 1.4235893
#> [13,] 1.4914569 1.450032 1.6247054 1.4920596 1.4806474 1.5839701 1.6188176
#> [14,] 1.2537564 1.208664 1.2219562 1.2492643 1.2451434 1.4184832 1.2740312
#> [15,] 1.2117198 1.197533 1.1892684 1.2103585 1.2043671 1.2365956 1.2591659
#> [16,] 1.1414471 1.137668 1.1241106 1.1360576 1.1440819 1.1523448 1.0931569
#>            [,8]      [,9]     [,10]
#>  [1,] 1.0215715 1.1045176 1.1116220
#>  [2,] 0.7205037 0.8418778 0.8501212
#>  [3,] 1.0399474 1.1603842 1.1567365
#>  [4,] 1.3158066 1.4271740 1.4399610
#>  [5,] 1.0479828 1.0123399 1.0156064
#>  [6,] 1.1145679 1.1678661 1.1733476
#>  [7,] 1.5770105 1.4959642 1.5102808
#>  [8,] 1.3905476 1.1823149 1.1862538
#>  [9,] 1.2793412 1.2046692 1.2139537
#> [10,] 1.1758852 1.2370983 1.2456182
#> [11,] 1.4289084 1.3807875 1.3819703
#> [12,] 1.2935091 1.3915455 1.4073177
#> [13,] 1.9939486 1.4888409 1.5011507
#> [14,] 1.1674317 1.2563327 1.2909260
#> [15,] 1.3057431 1.2090995 1.2080818
#> [16,] 0.9903576 1.1384504 1.1518946
# }