Posterior Frailty Variances

Function for computing the posterior frailty variances of the time-dependent shared frailty Cox model.

Recalling the structure of the frailty \(Z_{jk} = \alpha_j + \epsilon_{jk}, \forall j,k\) with \(k=1,\dots,L\) and \(j=1,\dots,N\) as being composed by the sum of two independent gamma distributions:

\(\alpha_j \sim gamma(\mu_1/\nu, 1/\nu), \forall j\)

\(\epsilon_{jk} \sim gamma(\mu_2/\gamma_k, 1/\gamma_k), \forall j,k\)

The posterior frailty variance is \(var(\hat{Z}_{jk}) = var(\hat{\alpha}_{j}/\hat{\alpha}_{max}) + var(\hat{\epsilon}_{jk}/\hat{\epsilon}_{max}\)). This function allows to get either the entire posterior frailty variance \(var(\hat{Z}_{jk})\) or its time-independent \(var(\frac{\hat{\alpha}_{j}}{\hat{\alpha}_{\text{max}}})\) or time-dependent \(var(\frac{\hat{\epsilon}_{jk}}{\hat{\epsilon}_{\text{max}}})\) components. The user can control which components to display using the flag_eps and flag_alpha parameters. Only one of these flags can be set to TRUE at a time.

Usage

post_frailty_var(object, flag_eps = FALSE, flag_alpha = FALSE)

Arguments

object

S3 object of class 'AdPaik' returned by the main model output, that contains all the information for the computation of the frailty standard deviation.

flag_eps

Logical flag indicating whether to extract only the time-dependent posterior frailty estimates. Default is FALSE.

flag_alpha

Logical flag indicating whether to extract only the time-independent posterior frailty estimates. Default is FALSE.

Value

Vector or matrix of posterior frailty variances, depending on the flag_eps and flag_alpha values. Specifically:

It is a vector of length equal to the N containing posterior frailty variances for \(\alpha_j, \forall j\). In this case the flag_eps must be FALSE and the flag_alpha must be TRUE.

Matrix of dimension (N, L) containing posterior frailty variances for \(\epsilon_{jk}, \forall j,k\). In this case the flag_eps must be TRUE and the flag_alpha must be FALSE.

Matrix of dimension (N, L) containing posterior frailty variances for \(Z_{jk} \forall j,k\). In this case the flag_eps must be FALSE and the flag_alpha must be FALSE.

Examples

# Consider the 'Academic Dropout dataset'
data(data_dropout)

# Define the variables needed for the model execution
formula <- time_to_event ~ Gender + CFUP + cluster(group)
time_axis <- c(1.0, 1.4, 1.8, 2.3, 3.1, 3.8, 4.3, 5.0, 5.5, 5.8, 6.0)
eps <- 1e-10
categories_range_min <- c(-8, -2, eps, eps, eps)
categories_range_max <- c(-eps, 0, 1 - eps, 1, 10)

# \donttest{
# Call the main model
result <- AdPaikModel(formula, data_dropout, time_axis,
                      categories_range_min, categories_range_max)

post_frailty_var(result)
#>              [,1]       [,2]       [,3]        [,4]        [,5]       [,6]
#>  [1,] 0.009208044 0.01948431 0.01625025 0.009189024 0.011346526 0.01675298
#>  [2,] 0.008539012 0.01817776 0.01315508 0.008519128 0.010797676 0.01996978
#>  [3,] 0.035183553 0.04655698 0.04659608 0.035128479 0.037542408 0.05106555
#>  [4,] 0.016448520 0.03043304 0.02706068 0.016410356 0.018914331 0.02753624
#>  [5,] 0.012632917 0.02382767 0.01908468 0.012583282 0.014817968 0.02400006
#>  [6,] 0.016924957 0.02747708 0.02728673 0.016921054 0.019343282 0.02653620
#>  [7,] 0.046995368 0.05971798 0.05902756 0.046932157 0.049193506 0.06076235
#>  [8,] 0.022012331 0.03307782 0.03135558 0.021937817 0.024142469 0.03523235
#>  [9,] 0.032894942 0.04480225 0.04406581 0.032843980 0.035311593 0.04788315
#> [10,] 0.008963862 0.01952388 0.01473205 0.008930184 0.010906923 0.01790924
#> [11,] 0.033919801 0.04516655 0.04967971 0.033862380 0.036118109 0.04895768
#> [12,] 0.008987794 0.01912156 0.01576454 0.009031558 0.011212061 0.02221939
#> [13,] 0.012685763 0.02151703 0.02026922 0.012672312 0.014740807 0.02456841
#> [14,] 0.007730267 0.01592522 0.01184412 0.007712705 0.009729408 0.01914157
#> [15,] 0.025736212 0.03710164 0.03615302 0.025683797 0.027971641 0.04124472
#> [16,] 0.016810824 0.02804826 0.02583515 0.016752497 0.019092400 0.03049165
#>             [,7]       [,8]        [,9]       [,10]
#>  [1,] 0.03837031 0.02367616 0.009189121 0.010241579
#>  [2,] 0.04191422 0.02535602 0.008509954 0.009576217
#>  [3,] 0.07379331 0.05902814 0.035117137 0.036137814
#>  [4,] 0.04806615 0.03201287 0.016422522 0.017508341
#>  [5,] 0.04594658 0.03762429 0.012581283 0.013622995
#>  [6,] 0.05096494 0.03837075 0.016909469 0.017969404
#>  [7,] 0.08491996 0.08396285 0.046904744 0.048009204
#>  [8,] 0.06460903 0.06305173 0.021941118 0.022991155
#>  [9,] 0.07086813 0.06905274 0.032843122 0.033922219
#> [10,] 0.05593610 0.02253700 0.008946375 0.010007231
#> [11,] 0.07109076 0.06700919 0.033871035 0.034914538
#> [12,] 0.04395023 0.02000231 0.009037670 0.010124093
#> [13,] 0.05654892 0.05180148 0.012670479 0.013750489
#> [14,] 0.04145595 0.01866158 0.007745254 0.008935478
#> [15,] 0.06927522 0.06061676 0.025676390 0.026702017
#> [16,] 0.05141058 0.03347117 0.016757903 0.017851735
# }